+++ src/transformers/visitors/toContractVisitor.ts
@@ -7,12 +7,7 @@ import { traverseNodesFast } from '../../traverse/traverse.js';
 import NodePath from '../../traverse/NodePath.js';
 import { VariableBinding } from '../../traverse/Binding.js';
 import { ContractDefinitionIndicator,FunctionDefinitionIndicator } from '../../traverse/Indicator.js';
-// below stub will only work with a small subtree - passing a whole AST will always give true!
-// useful for subtrees like ExpressionStatements
-const interactsWithSecretVisitor = (thisPath: NodePath, thisState: any) => {
-  if (thisPath.scope.getReferencedBinding(thisPath.node)?.isSecret)
-    thisState.interactsWithSecret = true;
-};
+import { interactsWithSecretVisitor } from './common.js';
 
 // here we find any public state variables which interact with secret states
 // and hence need to be included in the verification calculation
@@ -68,6 +63,15 @@ const findCustomInputsVisitor = (thisPath: NodePath, thisState: any) => {
 }
   }
 };
+
+
+function parentnewASTPointer(parent , path , newNode){
+  if (Array.isArray(parent._newASTPointer[path.containerName])) {
+    parent._newASTPointer[path.containerName].push(newNode);
+  } else {
+    parent._newASTPointer[path.containerName] = newNode;
+  }
+}
 
 let internalFuncInteractsWithSecret = false;
 /**
@@ -475,11 +479,7 @@ export default {
 
       const newNode = buildNode('BinaryOperation', { operator });
       node._newASTPointer = newNode;
-      if (Array.isArray(parent._newASTPointer[path.containerName])) {
-        parent._newASTPointer[path.containerName].push(newNode);
-      } else {
-        parent._newASTPointer[path.containerName] = newNode;
-      }
+      parentnewASTPointer(parent, path , newNode);
     },
   },
 
@@ -498,11 +498,7 @@ export default {
       const { operator, prefix, subExpression } = node;
       const newNode = buildNode(node.nodeType, { operator, prefix, subExpression });
       node._newASTPointer = newNode;
-      if (Array.isArray(parent._newASTPointer[path.containerName])) {
-        parent._newASTPointer[path.containerName].push(newNode);
-      } else {
-        parent._newASTPointer[path.containerName] = newNode;
-      }
+      parentnewASTPointer(parent, path , newNode);
     }
   },
 
@@ -609,11 +605,7 @@ export default {
       const newNode = buildNode('Identifier', { name });
 
       // node._newASTPointer = // no pointer needed, because this is a leaf, so we won't be recursing any further.
-      if (Array.isArray(parent._newASTPointer[path.containerName])) {
-        parent._newASTPointer[path.containerName].push(newNode);
-      } else {
-        parent._newASTPointer[path.containerName] = newNode;
-      }
+      parentnewASTPointer(parent, path , newNode);
     },
   },
 
@@ -630,12 +622,7 @@ export default {
       // node._newASTPointer = // no pointer needed, because this is a leaf, so we won't be recursing any further.
 
       const newNode = buildNode('Literal', { value, kind });
-
-      if (Array.isArray(parent._newASTPointer[path.containerName])) {
-        parent._newASTPointer[path.containerName].push(newNode);
-      } else {
-        parent._newASTPointer[path.containerName] = newNode;
-      }
+      parentnewASTPointer(parent, path , newNode);
     },
   },
 
@@ -653,12 +640,7 @@ export default {
         newNode = buildNode('MemberAccess', { memberName: node.memberName });
         node._newASTPointer = newNode;
       }
-
-      if (Array.isArray(parent._newASTPointer[path.containerName])) {
-        parent._newASTPointer[path.containerName].push(newNode);
-      } else {
-        parent._newASTPointer[path.containerName] = newNode;
-      }
+      parentnewASTPointer(parent, path , newNode);
     },
   },
 
@@ -668,11 +650,7 @@ export default {
 
       const newNode = buildNode('IndexAccess');
       node._newASTPointer = newNode;
-      if (Array.isArray(parent._newASTPointer[path.containerName])) {
-        parent._newASTPointer[path.containerName].push(newNode);
-      } else {
-        parent._newASTPointer[path.containerName] = newNode;
-      }
+      parentnewASTPointer(parent, path , newNode);
     },
   },
 
@@ -682,11 +660,7 @@ export default {
 
       const newNode = buildNode('Mapping');
       node._newASTPointer = newNode;
-      if (Array.isArray(parent._newASTPointer[path.containerName])) {
-        parent._newASTPointer[path.containerName].push(newNode);
-      } else {
-        parent._newASTPointer[path.containerName] = newNode;
-      }
+      parentnewASTPointer(parent, path , newNode);
     },
   },
 
@@ -709,12 +683,7 @@ export default {
         // For now, we'll copy these into Solidity:
         newNode = buildNode('FunctionCall');
         node._newASTPointer = newNode;
-        if (Array.isArray(parent._newASTPointer[path.containerName])) {
-          parent._newASTPointer[path.containerName].push(newNode);
-        } else {
-          parent._newASTPointer[path.containerName] = newNode;
-        }
-
+        parentnewASTPointer(parent, path , newNode);
         return;
       }
 
@@ -724,12 +693,7 @@ export default {
 
         newNode = buildNode('FunctionCall');
         node._newASTPointer = newNode;
-
-        if (Array.isArray(parent._newASTPointer[path.containerName])) {
-          parent._newASTPointer[path.containerName].push(newNode);
-        } else {
-          parent._newASTPointer[path.containerName] = newNode;
-        }
+        parentnewASTPointer(parent, path , newNode);
         return;
       }
       if (path.isInternalFunctionCall()) {
@@ -763,29 +727,17 @@ export default {
           parameters: state.fnParameters,
          });
          node._newASTPointer = newNode;
-         if (Array.isArray(parent._newASTPointer[path.containerName])) {
-           parent._newASTPointer[path.containerName].push(newNode);
-          } else {
-           parent._newASTPointer[path.containerName] = newNode;
-          }
+         parentnewASTPointer(parent, path , newNode);
          return;
         }
         newNode = buildNode('FunctionCall');
         node._newASTPointer = newNode;
-        if (Array.isArray(parent._newASTPointer[path.containerName])) {
-          parent._newASTPointer[path.containerName].push(newNode);
-        } else {
-          parent._newASTPointer[path.containerName] = newNode;
-        }
+        parentnewASTPointer(parent, path , newNode);
       }
       if (node.kind !== 'typeConversion') {
         newNode = buildNode('FunctionCall');
         node._newASTPointer = newNode;
-        if (Array.isArray(parent._newASTPointer[path.containerName])) {
-          parent._newASTPointer[path.containerName].push(newNode);
-        } else {
-          parent._newASTPointer[path.containerName] = newNode;
-        }
+        parentnewASTPointer(parent, path , newNode);
         state.skipSubNodes = true;
         return;
       }
@@ -793,10 +745,7 @@ export default {
         type: node.typeDescriptions.typeString,
       });
      node._newASTPointer = newNode;
-      if (Array.isArray(parent._newASTPointer[path.containerName])) {
-        parent._newASTPointer[path.containerName].push(newNode); }  else {
-        parent._newASTPointer[path.containerName] = newNode;
-      }
+     parentnewASTPointer(parent, path , newNode);
     },
   },
 }
